refactor: entity 생성 객체 분리
