+++ H-Mingle/src/main/java/com/hyundai/hmingle/support/OauthEntityGenerator.java
@@ -1,0 +1,11 @@
+package com.hyundai.hmingle.support;
+
+import org.springframework.http.HttpEntity;
+import org.springframework.util.MultiValueMap;
+
+public interface OauthEntityGenerator {
+
+	HttpEntity<MultiValueMap<String, String>> createForAccessToken(String redirectUrl, String authorizationCode);
+
+	HttpEntity<Void> createForUserInfo(String accessToken);
+}
